Warning fix
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/ptmalloc@24648 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/pooledscfclass.h
@@ -124,7 +124,7 @@ public:
         p.mti = mtiRegisterAlloc (n, typeid (Pool).name());
       }
       else
-        mtiUpdateAmount (p.mti, 1, n);
+        mtiUpdateAmount (p.mti, 1, (int)n);
     #endif
     }
     p.allocedEntries++;
